pyproject.toml
[project]
authors = [
    {name = "Alina Makarovskaya", email = "makarovskaya1307@gmail.com"},
]
requires-python = ">=3.11"
name = "hexlet-code"
version = "0.1.0"
description = "Project #3 - Page Analyzer"
readme = "README.md"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "validators>=0.34.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.0.1",
    "ruff>=0.9.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]
